Restored previous sources (#3.9.9.1)

+++ src/main/java/net/dzikoysk/funnyguilds/command/admin/AxcName.java
@@ -1,10 +1,16 @@
 package net.dzikoysk.funnyguilds.command.admin;
 
 import net.dzikoysk.funnyguilds.basic.Guild;
+import net.dzikoysk.funnyguilds.basic.Region;
 import net.dzikoysk.funnyguilds.basic.util.GuildUtils;
+import net.dzikoysk.funnyguilds.basic.util.RegionUtils;
 import net.dzikoysk.funnyguilds.command.util.Executor;
 import net.dzikoysk.funnyguilds.data.Manager;
 import net.dzikoysk.funnyguilds.data.Messages;
+import net.dzikoysk.funnyguilds.data.Settings;
+import net.dzikoysk.funnyguilds.data.database.DatabaseGuild;
+import net.dzikoysk.funnyguilds.data.database.DatabaseRegion;
+import net.dzikoysk.funnyguilds.data.flat.Flat;
 import net.dzikoysk.funnyguilds.util.StringUtils;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -13,11 +19,11 @@ public class AxcName implements Executor {
 
     @Override
     public void execute(CommandSender sender, String[] args) {
-        Messages messages = Messages.getInstance();
+        Messages m = Messages.getInstance();
         Player player = (Player) sender;
 
         if (!player.hasPermission("funnyguilds.admin")) {
-            player.sendMessage(messages.getMessage("permission"));
+            player.sendMessage(m.getMessage("permission"));
             return;
         }
 
@@ -38,23 +44,28 @@ public class AxcName implements Executor {
             return;
         }
 
-        //Settings settings = Settings.getInstance();
         Guild guild = GuildUtils.byTag(tag);
-        //Region region = guild.getRegion();
+        Region region = RegionUtils.get(guild.getRegion());
 
+        Settings s = Settings.getInstance();
         Manager.getInstance().stop();
-        /*if(settings.flat){
+        if (s.flat) {
             Flat.getGuildFile(guild).delete();
-			Flat.getRegionFile(region).delete();
-		}
-		if(settings.mysql){
-			new DatabaseGuild(guild).delete();
-			new DatabaseRegion(region).delete();
-		}*/
+            Flat.getRegionFile(region).delete();
+        }
+        if (s.mysql) {
+            new DatabaseGuild(guild).delete();
+            new DatabaseRegion(region).delete();
+        }
+        guild.getRegions().remove(region.getName());
         guild.setName(name);
+        region.setName(name);
+        guild.setRegion(name);
+        guild.getRegions().add(name);
         Manager.getInstance().start();
 
         player.sendMessage(StringUtils.colored("&7Zmieniono nazwe gildii na &b" + guild.getName() + "&7!"));
+        return;
     }
 
 }
